requirements.txt
Flask==1.1.2
google-api-python-client==1.9.3
numpy==1.18.1
Pillow==7.1.2
gunicorn==20.0.4